requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/requirements.in
#
pyserial==3.5
    # via
    #   -r requirements/requirements.in
    #   pyserial-asyncio
pyserial-asyncio==0.6
    # via -r requirements/requirements.in
